requirements.txt
Django==1.7.4
django-extensions==1.5.1
pyyaml==3.11
celery==3.1.17
